+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Utxo.hs
@@ -78,7 +78,6 @@ import Cardano.Ledger.Shelley.UTxO (UTxO (..))
 import Cardano.Ledger.ShelleyMA.Rules (ShelleyMAUtxoPredFailure)
 import qualified Cardano.Ledger.ShelleyMA.Rules as ShelleyMA
   ( validateOutsideValidityIntervalUTxO,
-    validateTriesToForgeADA,
     validateValueNotConservedUTxO,
   )
 import Cardano.Ledger.TxIn (TxIn)
@@ -402,9 +401,7 @@ utxoTransition = do
   runTest $
     ShelleyMA.validateValueNotConservedUTxO pp utxo stakepools txBody
 
-  {-   adaID ∉ supp mint tx   -}
-  runTestOnSignal $
-    ShelleyMA.validateTriesToForgeADA txBody
+  {-   adaID ∉ supp mint tx  - check not needed because mint field of type MultiAsset cannot contain ada  -}
 
   {-   ∀ txout ∈ allOuts txb, getValue txout ≥ inject (serSize txout ∗ coinsPerUTxOByte pp) -}
   runTest $ validateOutputTooSmallUTxO pp $ allSizedOuts txBody