+++ eras/shelley-ma/test-suite/src/Test/Cardano/Ledger/Mary/Golden.hs
@@ -20,7 +20,7 @@ where
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Core (hashScript)
 import Cardano.Ledger.Mary (MaryEra)
-import Cardano.Ledger.Mary.Value (AssetName (..), MaryValue (..), PolicyID (..))
+import Cardano.Ledger.Mary.Value (AssetName (..), MaryValue (..), MultiAsset (..), PolicyID (..))
 import Cardano.Ledger.ShelleyMA.Rules (scaledMinDeposit)
 import Cardano.Ledger.ShelleyMA.Timelocks (Timelock (..))
 import Cardano.Ledger.Slot (SlotNo (..))
@@ -84,107 +84,116 @@ goldenScaledMinDeposit =
     [ testCase "one policy, one (smallest) name" $
         scaledMinDeposit
           ( MaryValue 1407406 $
-              Map.singleton pid1 (Map.fromList [(smallestName, 1)])
+              MultiAsset $
+                Map.singleton pid1 (Map.fromList [(smallestName, 1)])
           )
           minUTxO
           @?= Coin 1407406,
       testCase "one policy, one (small) name" $
         scaledMinDeposit
           ( MaryValue 1444443 $
-              Map.singleton
-                pid1
-                (Map.fromList [(smallName 1, 1)])
+              MultiAsset $
+                Map.singleton
+                  pid1
+                  (Map.fromList [(smallName 1, 1)])
           )
           minUTxO
           @?= Coin 1444443,
       testCase "one policy, one (real) name" $
         scaledMinDeposit
           ( MaryValue 1444443 $
-              Map.singleton
-                pid1
-                (Map.fromList [(realName, 1)])
+              MultiAsset $
+                Map.singleton
+                  pid1
+                  (Map.fromList [(realName, 1)])
           )
           minUTxO
           @?= Coin 1481480,
       testCase "one policy, three (small) name" $
         scaledMinDeposit
           ( MaryValue 1555554 $
-              Map.singleton
-                pid1
-                ( Map.fromList
-                    [ (smallName 1, 1),
-                      (smallName 2, 1),
-                      (smallName 3, 1)
-                    ]
-                )
+              MultiAsset $
+                Map.singleton
+                  pid1
+                  ( Map.fromList
+                      [ (smallName 1, 1),
+                        (smallName 2, 1),
+                        (smallName 3, 1)
+                      ]
+                  )
           )
           minUTxO
           @?= Coin 1555554,
       testCase "one policy, one (largest) name" $
         scaledMinDeposit
           ( MaryValue 1555554 $
-              Map.singleton
-                pid1
-                (Map.fromList [(largestName 65, 1)])
+              MultiAsset $
+                Map.singleton
+                  pid1
+                  (Map.fromList [(largestName 65, 1)])
           )
           minUTxO
           @?= Coin 1555554,
       testCase "one policy, three (largest) name" $
         scaledMinDeposit
           ( MaryValue 1962961 $
-              Map.singleton
-                pid1
-                ( Map.fromList
-                    [ (largestName 65, 1),
-                      (largestName 66, 1),
-                      (largestName 67, 1)
-                    ]
-                )
+              MultiAsset $
+                Map.singleton
+                  pid1
+                  ( Map.fromList
+                      [ (largestName 65, 1),
+                        (largestName 66, 1),
+                        (largestName 67, 1)
+                      ]
+                  )
           )
           minUTxO
           @?= Coin 1962961,
       testCase "two policies, one (smallest) name" $
         scaledMinDeposit
           ( MaryValue 1592591 $
-              Map.fromList
-                [ ( pid1,
-                    Map.fromList [(smallestName, 1)]
-                  ),
-                  ( pid2,
-                    Map.fromList [(smallestName, 1)]
-                  )
-                ]
+              MultiAsset $
+                Map.fromList
+                  [ ( pid1,
+                      Map.fromList [(smallestName, 1)]
+                    ),
+                    ( pid2,
+                      Map.fromList [(smallestName, 1)]
+                    )
+                  ]
           )
           minUTxO
           @?= Coin 1592591,
       testCase "two policies, two (small) names" $
         scaledMinDeposit
           ( MaryValue 1629628 $
-              Map.fromList
-                [ ( pid1,
-                    Map.fromList [(smallName 1, 1)]
-                  ),
-                  ( pid2,
-                    Map.fromList [(smallName 2, 1)]
-                  )
-                ]
+              MultiAsset $
+                Map.fromList
+                  [ ( pid1,
+                      Map.fromList [(smallName 1, 1)]
+                    ),
+                    ( pid2,
+                      Map.fromList [(smallName 2, 1)]
+                    )
+                  ]
           )
           minUTxO
           @?= Coin 1629628,
       testCase "three policies, ninety-six (small) names" $
         scaledMinDeposit
           ( MaryValue 7407400 $
-              Map.fromList
-                [ ( pid1,
-                    Map.fromList $ map ((,1) . smallName) [32 .. 63]
-                  ),
-                  ( pid2,
-                    Map.fromList $ map ((,1) . smallName) [64 .. 95]
-                  ),
-                  ( pid3,
-                    Map.fromList $ map ((,1) . smallName) [96 .. 127]
-                  )
-                ]
+              MultiAsset $
+                Map.fromList
+                  [ ( pid1,
+                      Map.fromList $ map ((,1) . smallName) [32 .. 63]
+                    ),
+                    ( pid2,
+                      Map.fromList $ map ((,1) . smallName) [64 .. 95]
+                    ),
+                    ( pid3,
+                      Map.fromList $ map ((,1) . smallName) [96 .. 127]
+                    )
+                  ]
           )
           minUTxO
           @?= Coin 7407400