+++ libs/cardano-ledger-pretty/src/Cardano/Ledger/Pretty/Mary.hs
@@ -15,26 +15,35 @@ import Cardano.Ledger.Shelley.TxBody (ShelleyTxOut)
 import Cardano.Ledger.ShelleyMA.AuxiliaryData
 import Cardano.Ledger.ShelleyMA.Timelocks
 import Cardano.Ledger.ShelleyMA.TxBody
+import qualified Data.Map as Map
 import Prettyprinter (hsep, viaShow)
-
-ppValue :: MaryValue crypto -> PDoc
-ppValue v = case gettriples' v of
-  (n, triples, []) -> ppSexp "Value" [ppCoin (Coin n), ppList pptriple triples]
-  (n, triples, bad) -> ppSexp "Value" [ppCoin (Coin n), ppList pptriple triples, ppString "Bad " <> ppList ppPolicyID bad]
-  where
-    pptriple (i, asset, num) = hsep [ppPolicyID i, ppAssetName asset, ppInteger num]
 
 ppPolicyID :: PolicyID crypto -> PDoc
 ppPolicyID (PolicyID sh) = ppScriptHash sh
 
+instance PrettyA (PolicyID crypto) where prettyA x = ppSexp "PolicyID" [ppPolicyID x]
+
 ppAssetName :: AssetName -> PDoc
 ppAssetName = viaShow
 
-instance PrettyA (MaryValue crypto) where prettyA = ppValue
-
-instance PrettyA (PolicyID crypto) where prettyA x = ppSexp "PolicyID" [ppPolicyID x]
-
 instance PrettyA AssetName where prettyA x = ppSexp "AssetName" [ppAssetName x]
+
+ppMultiAsset :: MultiAsset crypto -> PDoc
+ppMultiAsset m = ppList pptriple (flattenMultiAsset m)
+  where
+    pptriple (i, asset, num) = hsep [ppPolicyID i, ppAssetName asset, ppInteger num]
+
+instance CC.Crypto crypto => PrettyA (MultiAsset crypto) where prettyA x = ppSexp "MultiAsset" [ppMultiAsset x]
+
+ppValue :: MaryValue crypto -> PDoc
+ppValue (MaryValue n m) = ppSexp "Value" $ [ppCoin (Coin n), ppMultiAsset m] ++ ppBad
+  where
+    ppBad = case getBadMultiAsset m of
+      [] -> []
+      bad -> [ppString "Bad " <> ppList ppPolicyID bad]
+    getBadMultiAsset (MultiAsset ma) = Map.keys (Map.filter Map.null ma)
+
+instance PrettyA (MaryValue crypto) where prettyA = ppValue
 
 ppTimelock :: CC.Crypto crypto => Timelock crypto -> PDoc
 ppTimelock (RequireSignature akh) =